add add Action command

+++ Structs.go
@@ -7,10 +7,11 @@ type Webhook struct {
 }
 
 //Action webhook struct
+//Mode (0 = script, 1 = action)
 type Action struct {
 	ID     int64  `db:"pkID" orm:"pk,ai"`
 	Name   string `db:"name"`
 	HookID string `db:"hookID"`
-	Mode   int16  `db:"mode"`
+	Mode   int8   `db:"mode"`
 	File   string `db:"file"`
 }
